requirements.txt
aiogram>=3.3.0
# Or if you prefer python-telegram-bot:
# python-telegram-bot>=20.7

mcp>=0.9.0
aiohttp>=3.9.0
apscheduler>=3.10.0
sqlalchemy>=2.0.0
alembic>=1.13.0
pydantic>=2.5.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
pytz>=2023.3
aiosqlite>=0.19.0
asyncpg>=0.29.0
redis>=5.0.0
